getting close to finilizing tree dir

+++ src/main/Gui.java
@@ -9,6 +9,7 @@ import java.awt.event.KeyEvent;
 import java.io.IOException;
 import java.io.File;
 import java.util.*;
+import main.FileTree;
 
 
 /**
@@ -46,7 +47,11 @@ public class Gui extends JFrame{
     // Attr added by Lucas not Intellji form designer
     private FileDialog fileChooser;
     private String projectPath;
-    private Queue<String> recentObfuscated;
+    private boolean isProject;
+    private JMenuItem delete;
+    private JMenuItem undo;
+    private JMenuItem redo;
+    //private Queue<String> recentObfuscated;
 
     public Gui() {
         super("LuaGuard");
@@ -67,13 +72,30 @@ public class Gui extends JFrame{
 
         // Adding edit menu tab with its menu items
         JMenu edit = new JMenu("Edit");
-        JMenuItem undo = new JMenuItem("Undo");
-        JMenuItem redo = new JMenuItem("Redo");
-        JMenuItem delete = new JMenuItem("Delete");
+        undo = new JMenuItem("Undo");
+        redo = new JMenuItem("Redo");
+        delete = new JMenuItem("Delete");
+        undo.setEnabled(false);
+        redo.setEnabled(false);
+        delete.setEnabled(false);
 
         // Init file chooser and project path for selecting files/dirs
         fileChooser = new FileDialog(this, "Choose a file", FileDialog.LOAD);
         projectPath = "";
+
+        // Init status pane with welcome/info message
+        statusTextPane.setText("Welcome to LuaGuard!\n" +
+                "Create/open a project to begin\n" +
+                "Or Start typing Lua code into the Lua Editor TextPane\n" +
+                "For further help see our user manual at...\n");
+
+        // Add to menu
+        file.add(newproj);
+        file.add(open);
+        file.add(openProj);
+        edit.add(undo);
+        edit.add(redo);
+        edit.add(delete);
 
         // Create a project, by either selecting a directory or creating a new directory
         newproj.setAccelerator(
@@ -82,39 +104,7 @@ public class Gui extends JFrame{
             public void actionPerformed(ActionEvent e) {
                 // Set to only open dirs
                 System.setProperty("apple.awt.fileDialogForDirectories", "true");
-                // Toggle File Chooser, get filename and dir of file
-                fileChooser.setVisible(true);
-                String fn = fileChooser.getFile();
-                String fn_loc = fileChooser.getDirectory();
-
-                // No File was Chosen
-                if (fn == null) {
-                    updateStatusPanel("Project not created, no folder specified or created");
-                } else {    // Get Path of Directory
-                    projectPath = fn_loc + fn;
-                    File file =  new File(fn_loc + fn);
-                    String[] files = file.list();
-                    // Check for Obfuscated Dir in Project File Path
-                    boolean containsObfDir =  false;
-                    for (int i = 0; i < files.length; i++) {
-                        if (files[i].equals("obfuscated")){
-                            if (new File(fn_loc + files[i]).isDirectory()) {
-                                containsObfDir = true;
-                                break;
-                            }
-                        }
-                    }
-                    if (!containsObfDir) {
-                        new File(fn_loc+fn+"/obfuscated").mkdir();
-                        String currText = statusTextPane.getText();
-                        currText += "Added obfuscation directory\nObfuscated code will be sent there\n";
-                        statusTextPane.setText(currText);
-                    } else {
-                        String currText = statusTextPane.getText();
-                        currText += "Found project...obfuscation sent to obfuscated dir\n";
-                        statusTextPane.setText(currText);
-                    }
-                }
+                openOrNewFileDirProj(0);
             }
         });
 
@@ -125,37 +115,7 @@ public class Gui extends JFrame{
             public void actionPerformed(ActionEvent e) {
                 // Set to open only files
                 System.setProperty("apple.awt.fileDialogForDirectories", "false");
-                fileChooser.setVisible(true);
-                String fn = fileChooser.getFile();
-                String fn_loc = fileChooser.getDirectory();
-
-                // No file was chosen
-                if (fn == null) {
-                    String currText = statusTextPane.getText();
-                    currText += "No file Chosen\n";
-                    statusTextPane.setText(currText);
-                } else {
-                // File was chosen, check file extension
-                    projectPath = fn_loc + fn;
-                    String file_extension = fn.split("\\.")[1];
-                    if (!file_extension.equals("lua")) {
-                        String currText = statusTextPane.getText();
-                        currText += "Incompatible file type chosen, .lua files only supported\n";
-                        statusTextPane.setText(currText);
-                    } else {
-                        // Write text from file to luaEditorPane
-                        String tmp = fn_loc + fn;
-                        try {
-                            File file = new File(tmp);
-                            luaEditorPane.setPage(file.toURI().toURL());
-                        } catch (IOException ex) {
-                            // Catch exception if file not found
-                            String currText = statusTextPane.getText();
-                            currText += "Could not find file!\n";
-                            statusTextPane.setText(currText);
-                        }
-                    }
-                }
+                openOrNewFileDirProj(1);
             }
         });
 
@@ -164,37 +124,9 @@ public class Gui extends JFrame{
                 KeyStroke.getKeyStroke(KeyEvent.VK_O, (java.awt.event.InputEvent.SHIFT_MASK | (Toolkit.getDefaultToolkit().getMenuShortcutKeyMask()))));
         openProj.addActionListener(new ActionListener(  ) {
             public void actionPerformed(ActionEvent e) {
-                // System.out.println("Open Project...");
                 // Set to open only dirs
                 System.setProperty("apple.awt.fileDialogForDirectories", "true");
-                fileChooser.setVisible(true);
-                String fn = fileChooser.getFile();
-                String fn_loc = fileChooser.getDirectory() + fn;
-                if (fn == null) {
-                    System.out.println("You cancelled the choice");
-                } else {
-                    projectPath = fn_loc + fn;
-                    File file =  new File(fn_loc);
-                    String[] files = file.list();
-                    boolean containsObfDir =  false;
-                    for (int i = 0; i < files.length; i++) {
-                        if (files[i].equals("obfuscated")){
-                            if (new File(fn_loc + files[i]).isDirectory()) {
-                                containsObfDir = true;
-                                break;
-                            }
-                        }
-                    }
-                    if (!containsObfDir) {
-                        new File(fn_loc+"/obfuscated").mkdir();
-                    } else {
-                        String currText = statusTextPane.getText();
-                        currText += "Found project...obfuscation sent to obfuscated directory\n";
-                        statusTextPane.setText(currText);
-                    }
-
-
-                }
+                openOrNewFileDirProj(2);
             }
         });
 
@@ -202,9 +134,14 @@ public class Gui extends JFrame{
                 KeyStroke.getKeyStroke(KeyEvent.VK_D, (Toolkit.getDefaultToolkit().getMenuShortcutKeyMask())));
         delete.addActionListener(new ActionListener(  ) {
             public void actionPerformed(ActionEvent e) {
-                System.out.println("Delete This Project...");
-                File file = new File(projectPath);
-                file.delete();
+                if (isProject) {
+                    File file = new File(projectPath+"/obfuscated");
+                    file.delete();
+                } else {
+                    System.out.println("DOTHisLater");
+                    //File file = new File(projectPath);
+                    //file.delete();
+                }
             }
         });
 
@@ -212,6 +149,7 @@ public class Gui extends JFrame{
                 KeyStroke.getKeyStroke(KeyEvent.VK_Z, (Toolkit.getDefaultToolkit().getMenuShortcutKeyMask())));
         undo.addActionListener(new ActionListener(  ) {
             public void actionPerformed(ActionEvent e) {
+                redo.setEnabled(true);
                 System.out.println("undo recent obfuscated file...");
             }
         });
@@ -225,23 +163,13 @@ public class Gui extends JFrame{
             }
         });
 
-        //
-        file.add(newproj);
-        file.add(open);
-        file.add(openProj);
-        edit.add(undo);
-        edit.add(redo);
-        edit.add(delete);
-
-        statusTextPane.setText("Welcome to LuaGuard!\n" +
-                "Create/open a project to begin\n" +
-                "Or Start typing Lua code into the Lua Editor TextPane\n" +
-                "For further help see our user manual at...\n");
 
 
         obfuscateButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent actionEvent) {
+                delete.setEnabled(true);
+                undo.setEnabled(true);
                 // Get Lua Code
                 String luaCode = luaEditorPane.getText();
 
@@ -293,26 +221,85 @@ public class Gui extends JFrame{
 
     public void updateStatusPanel(String addText) {
         String currText = statusTextPane.getText();
-        currText += addText + "\n";
+        currText += addText;
         statusTextPane.setText(currText);
     }
 
-    public boolean checkForObfuscatedDir(String projectPath) {
+    public boolean checkForObfuscatedDir() {
         // Create boolean and init File objects to find sub dir and files
         boolean containsObfDir = false;
         File file =  new File(projectPath);
         String[] files = file.list();
+
         // Iterate through sub-dirs and files looking for obfuscated dir
         for (int i = 0; i < files.length; i++) {
             if (files[i].equals("obfuscated")){     //Found file "obfuscate"
                 // Check if Obfuscated file is a dir
-                if (new File(projectPath + files[i]).isDirectory()) {
+                if (new File(projectPath + "/" + files[i]).isDirectory()) {
                     containsObfDir = true;
                     break;
                 }
             }
         }
         return containsObfDir;
+    }
+
+    public void openOrNewFileDirProj(int type){
+        // int type => Type of Action
+        // 0 = New project
+        // 1 = Open file
+        // 2 = Open a Project
+        fileChooser.setVisible(true);
+        String fn = fileChooser.getFile();
+        String fn_loc = fileChooser.getDirectory();
+        if (fn == null) {
+            return;
+        } else {
+            projectPath = fn_loc + fn;
+            // Opening a lua file
+            if (type == 1) {
+                String file_extension = fn.split("\\.")[1];
+                if (!file_extension.equals("lua")) {
+                    updateStatusPanel("Incompatible file type chosen, .lua files only supported\n");
+                } else {
+                    // Write text from file to luaEditorPane
+                    try {
+                        File file = new File(projectPath);
+                        luaEditorPane.setPage(file.toURI().toURL());
+                    } catch (IOException ex) {
+                        // Catch exception if file not found
+                        updateStatusPanel("Could not find file!\n");
+                    }
+                }
+            }
+            // Creating a new project/Opening a project
+            if (type == 0 || type == 2) {
+                // Enabled to delete obfuscated directory
+                delete.setEnabled(true);
+                isProject = true;
+                // Check for Obfuscated Dir in Project File Path
+                boolean findObfDir = checkForObfuscatedDir();
+                if (!findObfDir) {
+                    new File(projectPath+"/obfuscated").mkdir();
+                    updateStatusPanel("Added obfuscation directory\nObfuscated code will be sent there\n");
+                } else {
+                    updateStatusPanel("Found project...obfuscation sent to obfuscated directory\n");
+                }
+                Container cp = getContentPane();
+                //cp.remove(projectDirectoryPanel);
+                //cp.add(new FileTree(new File(projectPath)));
+
+                projectDirectoryPanel.removeAll();
+                projectDirectoryPanel.updateUI();
+                JPanel tmp  = new FileTree(new File(projectPath));
+                projectDirectoryPanel = tmp;
+                projectDirectoryPanel.updateUI();
+                //projectDirectoryPanel = new FileTree(new File(projectPath));
+                //projectDirectoryPanel.updateUI();
+                //projectDirectoryPanel.
+
+            }
+        }
     }
 
 }
